chore: move useLayoutBgClassName back
