+++ libs/webb-ui-components/src/next-utils/index.ts
@@ -1,2 +1,2 @@
 export * from './sidebarState';
-export * from './temp';
+export * from './useLayoutBgClassName';
